Cleanup platform specific code
